+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Removed `Awesomite\StackTrace\Arguments\Values\MultipleValues`
 * Added tests for HHVM with flag `hhvm.php7.all=1`
+* `Awesomite\StackTrace\StackTrace::unserialize` throws `Awesomite\StackTrace\Exceptions\LogicException` instead of `LogicException`
 
 ## 0.9.2